IDEMPIERE-4601 Move Plugin manager from About to a Form
